Removed base class that was no longer common
