+++ src/postgres/specification-result-sql.ts
@@ -2,16 +2,17 @@ import {
     ChildProjections,
     ElementProjection,
     FactReference,
+    Label,
     Match,
+    PathCondition,
     ResultProjection,
     SingularProjection,
     Specification,
     SpecificationProjection,
 } from "jinaga";
 
-import { FactTypeMap, getFactTypeId, RoleMap } from "./maps";
+import { FactTypeMap, getFactTypeId, getRoleId, RoleMap } from "./maps";
 import { FactDescription, InputDescription, QueryDescription, SpecificationSqlQuery } from "./query-description";
-import { QueryDescriptionBuilder } from "./query-description-builder";
 
 export type FactByIdentifier = {
     [identifier: string]: FactDescription;
@@ -206,13 +207,11 @@ interface NamedResultComposer {
     resultComposer: ResultComposer;
 }
 
-class ResultDescriptionBuilder extends QueryDescriptionBuilder {
+class ResultDescriptionBuilder {
     constructor(
-        factTypes: FactTypeMap,
-        roleMap: RoleMap
-    ) {
-        super(factTypes, roleMap);
-    }
+        private factTypes: FactTypeMap,
+        private roleMap: RoleMap
+    ) { }
 
     buildDescription(start: FactReference[], specification: Specification): ResultDescription {
         // Verify that the number of start facts equals the number of inputs
@@ -332,6 +331,110 @@ class ResultDescriptionBuilder extends QueryDescriptionBuilder {
             queryDescription,
             knownFacts
         };
+    }
+
+    private addPathCondition(queryDescription: QueryDescription, knownFacts: FactByIdentifier, path: number[], unknown: Label, prefix: string, condition: PathCondition): { queryDescription: QueryDescription, knownFacts: FactByIdentifier } {
+        // If no input parameter has been allocated, allocate one now.
+        const input = queryDescription.inputByLabel(condition.labelRight);
+        if (input && input.factTypeParameter === 0) {
+            queryDescription = queryDescription.withInputParameter(input.label);
+        }
+
+        // Determine whether we have already written the output.
+        const knownFact = knownFacts[unknown.name];
+        const roleCount = condition.rolesLeft.length + condition.rolesRight.length;
+
+        // Walk up the right-hand side.
+        // This generates predecessor joins from a given or prior label.
+        let fact = knownFacts[condition.labelRight];
+        if (!fact) {
+            throw new Error(`Label ${condition.labelRight} not found. Known labels: ${Object.keys(knownFacts).join(", ")}`);
+        }
+        let type = fact.type;
+        let factIndex = fact.factIndex;
+        for (const [i, role] of condition.rolesRight.entries()) {
+            // If the type or role is not known, then no facts matching the condition can
+            // exist. The query is unsatisfiable.
+            const typeId = getFactTypeId(this.factTypes, type);
+            if (!typeId) {
+                return { queryDescription: QueryDescription.unsatisfiable, knownFacts };
+            }
+            const roleId = getRoleId(this.roleMap, typeId, role.name);
+            if (!roleId) {
+                return { queryDescription: QueryDescription.unsatisfiable, knownFacts };
+            }
+
+            const { query: queryWithParameter, parameterIndex: roleParameter } = queryDescription.withParameter(roleId);
+            if (i === roleCount - 1 && knownFact) {
+                // If we have already written the output, we can use the fact index.
+                queryDescription = queryWithParameter.withEdge(knownFact.factIndex, factIndex, roleParameter, path);
+                factIndex = knownFact.factIndex;
+            }
+            else {
+                // If we have not written the fact, we need to write it now.
+                const { query, factIndex: predecessorFactIndex } = queryWithParameter.withFact(role.targetType);
+                queryDescription = query.withEdge(predecessorFactIndex, factIndex, roleParameter, path);
+                factIndex = predecessorFactIndex;
+            }
+            type = role.targetType;
+        }
+
+        const rightType = type;
+
+        // Walk up the left-hand side.
+        // We will need to reverse this walk to generate successor joins.
+        type = unknown.type;
+        const newEdges: {
+            roleId: number;
+            declaringType: string;
+        }[] = [];
+        for (const role of condition.rolesLeft) {
+            // If the type or role is not known, then no facts matching the condition can
+            // exist. The query is unsatisfiable.
+            const typeId = getFactTypeId(this.factTypes, type);
+            if (!typeId) {
+                return { queryDescription: QueryDescription.unsatisfiable, knownFacts };
+            }
+            const roleId = getRoleId(this.roleMap, typeId, role.name);
+            if (!roleId) {
+                return { queryDescription: QueryDescription.unsatisfiable, knownFacts };
+            }
+
+            newEdges.push({
+                roleId,
+                declaringType: type
+            });
+            type = role.targetType;
+        }
+
+        if (type !== rightType) {
+            throw new Error(`Type mismatch: ${type} is compared to ${rightType}`);
+        }
+
+        newEdges.reverse().forEach(({ roleId, declaringType }, i) => {
+            const { query: queryWithParameter, parameterIndex: roleParameter } = queryDescription.withParameter(roleId);
+            if (condition.rolesRight.length + i === roleCount - 1 && knownFact) {
+                queryDescription = queryWithParameter.withEdge(factIndex, knownFact.factIndex, roleParameter, path);
+                factIndex = knownFact.factIndex;
+            }
+            else {
+                const { query: queryWithFact, factIndex: successorFactIndex } = queryWithParameter.withFact(declaringType);
+                queryDescription = queryWithFact.withEdge(factIndex, successorFactIndex, roleParameter, path);
+                factIndex = successorFactIndex;
+            }
+        });
+
+        // If we have not captured the known fact, add it now.
+        if (!knownFact) {
+            knownFacts = { ...knownFacts, [unknown.name]: { factIndex, type: unknown.type } };
+            // If we have not written the output, write it now.
+            // Only write the output if we are not inside of an existential condition.
+            // Use the prefix, which will be set for projections.
+            if (path.length === 0) {
+                queryDescription = queryDescription.withOutput(prefix + unknown.name, unknown.type, factIndex);
+            }
+        }
+        return { queryDescription, knownFacts };
     }
 }
